Add Rustler under NIF category
I don't think this is strictly allowed, since it's not actually written in elixir, and is nor hex installable. I would argue it's relevant regardless, as it provides specific functionality in order to make Nifs easy and fast to write and use.

If you don't agree, feel free to decline the pull.