Update GitHub issue and pull request templates
At the moment we have 2.7k open issues that should be probably closed,
as the vast majority of them are out of date.

A few people have also been opening issues instead of pull requests, so
I think it's time to do something about it. I've updated the templates
to make it clear that new contributors can simply open a PR.